requirements.txt
Flask==3.0.3
numpy==1.26.4
opencv-python==4.9.0.80
Werkzeug==3.1.3
fastapi==0.104.1
uvicorn==0.28.0
ultralytics
psycopg2-binary
prometheus-client==0.20.0
hydra-core==1.3.2
shapely==2.0.3
cython_bbox==0.1.5
python-multipart
tensorrt==10.1.0